pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ctgov-search"
version = "5.1.0"
description = "Comprehensive Python toolkit for ClinicalTrials.gov search strategy validation, recall analysis, and multi-registry trial searching with ML optimization"
readme = "README.md"
license = {text = "MIT License - For Research Purposes"}
requires-python = ">=3.9"
authors = [
    {name = "CT.gov Search Strategy Project", email = "research@example.com"}
]
maintainers = [
    {name = "CT.gov Search Strategy Project", email = "research@example.com"}
]
keywords = [
    "clinical-trials",
    "clinicaltrials.gov",
    "systematic-review",
    "search-strategy",
    "recall-validation",
    "cochrane",
    "meta-analysis",
    "ICTRP",
    "MeSH",
    "RCT"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "requests>=2.28.0,<3.0",
    "psycopg2-binary>=2.9.0,<3.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0,<4.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "scipy>=1.10.0,<2.0",
    "pandas>=1.5.0,<3.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "safety>=2.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.0.0",
]
all = [
    "ctgov-search[dev,docs]",
]

[project.urls]
Homepage = "https://github.com/ctgov-search-strategies/ctgov-search"
Documentation = "https://github.com/ctgov-search-strategies/ctgov-search#readme"
Repository = "https://github.com/ctgov-search-strategies/ctgov-search.git"
Issues = "https://github.com/ctgov-search-strategies/ctgov-search/issues"

[project.scripts]
ctgov-search = "ctgov_search:main"
ctgov-workflow = "ctgov_workflow:main"
ctgov-validate = "aact_validation:main"

[tool.setuptools]
packages = ["ctgov_search", "scripts", "tests"]
py-modules = [
    "ctgov_search",
    "ctgov_config",
    "ctgov_utils",
    "ctgov_terms",
    "ctgov_workflow",
    "ctgov_fast",
    "ctgov_advanced",
    "aact_search",
    "aact_validation",
    "aact_full_validation",
    "aact_debug",
    "combined_search",
    "comprehensive_validation",
    "deep_optimizer",
    "direct_nct_search",
    "enhanced_strategy",
    "esc_cardiology_search",
    "expanded_validation",
    "max_recall_strategy",
    "multi_source_search",
    "scenario_optimizer",
    "challenging_conditions",
    "euctr_search",
    "anzctr_search",
    "relative_recall",
]

[tool.setuptools.package-data]
"*" = ["*.json", "*.csv", "*.txt"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=. --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "requires_aact: mark test as requiring AACT database credentials",
    "slow: mark test as slow running",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.flake8]
max-line-length = 120
exclude = [".git", "__pycache__", "build", "dist", "*.egg-info"]
ignore = ["E203", "W503"]

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true
